reduced logging + fixes for higher iters

+++ segment.h
@@ -23,7 +23,7 @@ typedef struct data {
 } data_t;
 
 data_t* segment_init(problem_t*, config_t*, segment_t*);
-int segment_burn(data_t*, int);
+int segment_burn(data_t*, int64_t);
 void segment_finalize(data_t*);
 void print_segment_blocks(data_t*);
 void print_smallest_mod(data_t* data, uint64_t mod);
